pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG pipeline for enterprise document analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "chromadb>=0.4.0",
    "tiktoken>=0.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "fpdf2>=2.7.0",
    "diagrams>=0.25.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.1.0", "pytest-asyncio>=0.23.0"]
vision = ["transformers>=4.40.0", "torch>=2.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentic_rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py313"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
